Pyxis now uses a combination of msgpack and lmdb.
The change to use msgpack allows the user to store an arbitrary number of tensors (numpy.ndarray) of different shapes and data types. Unfortunately, this means that several functions have been deprecated in favour of new ones. Please have a look in the Python notebooks to see what has changed.